提出Cursor的JS代码
