+++ package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "gsap": "^3.12.7",
     "sass-embedded": "^1.83.4",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
